refactor chatwitheino middleware helpers
