+++ quickstart/chatwitheino/cmd/ch07/main.go
@@ -39,13 +39,13 @@ import (
 	"github.com/cloudwego/eino/adk/prebuilt/deep"
 	"github.com/cloudwego/eino/callbacks"
 	"github.com/cloudwego/eino/components/tool"
-	"github.com/cloudwego/eino/compose"
 	"github.com/cloudwego/eino/schema"
 	"github.com/coze-dev/cozeloop-go"
 
 	adkstore "github.com/cloudwego/eino-examples/adk/common/store"
 	commontool "github.com/cloudwego/eino-examples/adk/common/tool"
 	"github.com/cloudwego/eino-examples/quickstart/chatwitheino/chatmodel"
+	"github.com/cloudwego/eino-examples/quickstart/chatwitheino/helpers"
 	"github.com/cloudwego/eino-examples/quickstart/chatwitheino/mem"
 	"github.com/cloudwego/eino-examples/quickstart/chatwitheino/msgops"
 )
@@ -138,19 +138,10 @@ Always use absolute paths when calling filesystem tools.`, projectRoot, projectR
 		MaxIteration:   50,
 		Handlers: []adk.TypedChatModelAgentMiddleware[M]{
 			newApprovalMiddleware[M](),
-			newSafeToolMiddleware[M](),
+			helpers.NewSafeToolMiddleware[M](),
 		},
 	}
-	if msgops.KindOf[M]() == msgops.KindMessage {
-		cfg.ModelRetryConfig = &adk.TypedModelRetryConfig[M]{
-			MaxRetries: 5,
-			IsRetryAble: func(_ context.Context, err error) bool {
-				return strings.Contains(err.Error(), "429") ||
-					strings.Contains(err.Error(), "Too Many Requests") ||
-					strings.Contains(err.Error(), "qpm limit")
-			},
-		}
-	}
+	helpers.ApplyMessageModelRetry(cfg)
 	agent, err := deep.NewTyped[M](ctx, cfg)
 	if err != nil {
 		_, _ = fmt.Fprintln(os.Stderr, err)
@@ -248,12 +239,6 @@ func newApprovalMiddleware[M adk.MessageType]() adk.TypedChatModelAgentMiddlewar
 	}
 }
 
-func newSafeToolMiddleware[M adk.MessageType]() adk.TypedChatModelAgentMiddleware[M] {
-	return &safeToolMiddleware[M]{
-		TypedBaseChatModelAgentMiddleware: &adk.TypedBaseChatModelAgentMiddleware[M]{},
-	}
-}
-
 func (m *approvalMiddleware[M]) WrapInvokableToolCall(
 	_ context.Context,
 	endpoint adk.InvokableToolCallEndpoint,
@@ -317,9 +302,9 @@ func (m *approvalMiddleware[M]) WrapStreamableToolCall(
 				return endpoint(ctx, storedArgs, opts...)
 			}
 			if data.DisapproveReason != nil {
-				return singleChunkReader(fmt.Sprintf("tool '%s' disapproved: %s", tCtx.Name, *data.DisapproveReason)), nil
+				return helpers.SingleChunkReader(fmt.Sprintf("tool '%s' disapproved: %s", tCtx.Name, *data.DisapproveReason)), nil
 			}
-			return singleChunkReader(fmt.Sprintf("tool '%s' disapproved", tCtx.Name)), nil
+			return helpers.SingleChunkReader(fmt.Sprintf("tool '%s' disapproved", tCtx.Name)), nil
 		}
 
 		isTarget, _, _ = tool.GetResumeContext[any](ctx)
@@ -332,70 +317,6 @@ func (m *approvalMiddleware[M]) WrapStreamableToolCall(
 
 		return endpoint(ctx, storedArgs, opts...)
 	}, nil
-}
-
-type safeToolMiddleware[M adk.MessageType] struct {
-	*adk.TypedBaseChatModelAgentMiddleware[M]
-}
-
-func (m *safeToolMiddleware[M]) WrapInvokableToolCall(
-	_ context.Context,
-	endpoint adk.InvokableToolCallEndpoint,
-	_ *adk.ToolContext,
-) (adk.InvokableToolCallEndpoint, error) {
-	return func(ctx context.Context, args string, opts ...tool.Option) (string, error) {
-		result, err := endpoint(ctx, args, opts...)
-		if err != nil {
-			if _, ok := compose.IsInterruptRerunError(err); ok {
-				return "", err
-			}
-			return fmt.Sprintf("[tool error] %v", err), nil
-		}
-		return result, nil
-	}, nil
-}
-
-func (m *safeToolMiddleware[M]) WrapStreamableToolCall(
-	_ context.Context,
-	endpoint adk.StreamableToolCallEndpoint,
-	_ *adk.ToolContext,
-) (adk.StreamableToolCallEndpoint, error) {
-	return func(ctx context.Context, args string, opts ...tool.Option) (*schema.StreamReader[string], error) {
-		sr, err := endpoint(ctx, args, opts...)
-		if err != nil {
-			if _, ok := compose.IsInterruptRerunError(err); ok {
-				return nil, err
-			}
-			return singleChunkReader(fmt.Sprintf("[tool error] %v", err)), nil
-		}
-		return safeWrapReader(sr), nil
-	}, nil
-}
-
-func singleChunkReader(msg string) *schema.StreamReader[string] {
-	r, w := schema.Pipe[string](1)
-	_ = w.Send(msg, nil)
-	w.Close()
-	return r
-}
-
-func safeWrapReader(sr *schema.StreamReader[string]) *schema.StreamReader[string] {
-	r, w := schema.Pipe[string](64)
-	go func() {
-		defer w.Close()
-		for {
-			chunk, err := sr.Recv()
-			if errors.Is(err, io.EOF) {
-				return
-			}
-			if err != nil {
-				_ = w.Send(fmt.Sprintf("\n[tool error] %v", err), nil)
-				return
-			}
-			_ = w.Send(chunk, nil)
-		}
-	}()
-	return r
 }
 
 func printAndCollectAssistantFromEvents[M adk.MessageType](events *adk.AsyncIterator[*adk.TypedAgentEvent[M]]) (string, *adk.InterruptInfo, error) {
