+++ quickstart/chatwitheino/helpers/middleware.go
@@ -1,0 +1,106 @@
+/*
+ * Copyright 2026 CloudWeGo Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package helpers
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"io"
+
+	"github.com/cloudwego/eino/adk"
+	"github.com/cloudwego/eino/components/tool"
+	"github.com/cloudwego/eino/compose"
+	"github.com/cloudwego/eino/schema"
+)
+
+// NewSafeToolMiddleware converts tool errors into error-message strings so that
+// a non-zero exit code or mid-stream failure is returned to the model as a
+// readable tool result instead of aborting the agent pipeline.
+func NewSafeToolMiddleware[M adk.MessageType]() adk.TypedChatModelAgentMiddleware[M] {
+	return &safeToolMiddleware[M]{
+		TypedBaseChatModelAgentMiddleware: &adk.TypedBaseChatModelAgentMiddleware[M]{},
+	}
+}
+
+type safeToolMiddleware[M adk.MessageType] struct {
+	*adk.TypedBaseChatModelAgentMiddleware[M]
+}
+
+func (m *safeToolMiddleware[M]) WrapInvokableToolCall(
+	_ context.Context,
+	endpoint adk.InvokableToolCallEndpoint,
+	_ *adk.ToolContext,
+) (adk.InvokableToolCallEndpoint, error) {
+	return func(ctx context.Context, args string, opts ...tool.Option) (string, error) {
+		result, err := endpoint(ctx, args, opts...)
+		if err != nil {
+			if _, ok := compose.IsInterruptRerunError(err); ok {
+				return "", err
+			}
+			return fmt.Sprintf("[tool error] %v", err), nil
+		}
+		return result, nil
+	}, nil
+}
+
+func (m *safeToolMiddleware[M]) WrapStreamableToolCall(
+	_ context.Context,
+	endpoint adk.StreamableToolCallEndpoint,
+	_ *adk.ToolContext,
+) (adk.StreamableToolCallEndpoint, error) {
+	return func(ctx context.Context, args string, opts ...tool.Option) (*schema.StreamReader[string], error) {
+		sr, err := endpoint(ctx, args, opts...)
+		if err != nil {
+			if _, ok := compose.IsInterruptRerunError(err); ok {
+				return nil, err
+			}
+			return SingleChunkReader(fmt.Sprintf("[tool error] %v", err)), nil
+		}
+		return safeWrapReader(sr), nil
+	}, nil
+}
+
+// SingleChunkReader returns a StreamReader that emits one string then EOF.
+func SingleChunkReader(msg string) *schema.StreamReader[string] {
+	r, w := schema.Pipe[string](1)
+	_ = w.Send(msg, nil)
+	w.Close()
+	return r
+}
+
+// safeWrapReader proxies chunks from sr; on a stream error it emits the error
+// as a final chunk instead of propagating it, so the model sees a complete
+// (if error-annotated) tool result rather than a pipeline failure.
+func safeWrapReader(sr *schema.StreamReader[string]) *schema.StreamReader[string] {
+	r, w := schema.Pipe[string](64)
+	go func() {
+		defer w.Close()
+		for {
+			chunk, err := sr.Recv()
+			if errors.Is(err, io.EOF) {
+				return
+			}
+			if err != nil {
+				_ = w.Send(fmt.Sprintf("\n[tool error] %v", err), nil)
+				return
+			}
+			_ = w.Send(chunk, nil)
+		}
+	}()
+	return r
+}
